refactor: move cache modules to providers
