+++ apps/backend/app/domains/admin/application/jobs_service.py
@@ -5,10 +5,10 @@ import json
 from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.core.cache import cache as shared_cache
 from app.core.config import settings
-from app.core.redis_utils import create_async_redis
 from app.models.background_job_history import BackgroundJobHistory
+from app.providers.cache import cache as shared_cache
+from app.providers.redis_utils import create_async_redis
 
 
 class JobsService:
