+++ tests/unit/test_health_readyz.py
@@ -12,7 +12,7 @@ os.environ.setdefault("TESTING", "True")
 sys.modules.setdefault("app", importlib.import_module("apps.backend.app"))
 
 from app.api import health as health_module  # noqa: E402
-from app.core.cache import cache as shared_cache  # noqa: E402
+from app.providers.cache import cache as shared_cache  # noqa: E402
 from app.providers.db.session import get_db  # noqa: E402
 
 app = FastAPI()
